+++ app/src/pages/login/LoginPage.tsx
@@ -1,13 +1,11 @@
-import React, { useCallback } from "react";
-import { Controller, useForm } from "react-hook-form";
-import { graphql, useMutation } from "react-relay";
+import React from "react";
 import { css } from "@emotion/react";
 
-import { Button, Flex, Form, TextField, View } from "@arizeai/components";
+import { Flex, View } from "@arizeai/components";
 
 import { Logo } from "@phoenix/components/nav/Logo";
 
-import type { LoginPageMutation } from "./__generated__/LoginPageMutation.graphql";
+import { LoginForm } from "./LoginForm";
 
 export function LoginPage() {
   return (
@@ -35,89 +33,5 @@ export function LoginPage() {
         <LoginForm />
       </View>
     </main>
-  );
-}
-
-type LoginFormParams = {
-  email: string;
-  password: string;
-};
-
-export function LoginForm() {
-  const [commit, isCommiting] = useMutation<LoginPageMutation>(graphql`
-    mutation LoginPageMutation($email: String!, $password: String!) {
-      login(input: { email: $email, password: $password }) {
-        success
-      }
-    }
-  `);
-  const onSubmit = useCallback(
-    (params: LoginFormParams) => {
-      commit({
-        variables: params,
-        onCompleted: (response) => {
-          if (response.login.success) {
-            return;
-          } else {
-            return;
-          }
-        },
-      });
-    },
-    [commit]
-  );
-  const { control, handleSubmit } = useForm<LoginFormParams>({
-    defaultValues: { email: "", password: "" },
-  });
-  return (
-    <Form>
-      <Controller
-        name="email"
-        control={control}
-        render={({ field: { onChange, value } }) => (
-          <TextField
-            label="Email"
-            name="email"
-            isRequired
-            type="email"
-            onChange={onChange}
-            value={value}
-            placeholder="your email address"
-          />
-        )}
-      />
-      <Controller
-        name="password"
-        control={control}
-        render={({ field: { onChange, value } }) => (
-          <TextField
-            label="Password"
-            name="password"
-            type="password"
-            isRequired
-            onChange={onChange}
-            value={value}
-            placeholder="your password"
-          />
-        )}
-      />
-      <div
-        css={css`
-          margin-top: var(--ac-global-dimension-size-400);
-          margin-bottom: var(--ac-global-dimension-size-50);
-          button {
-            width: 100%;
-          }
-        `}
-      >
-        <Button
-          variant="primary"
-          loading={isCommiting}
-          onClick={handleSubmit(onSubmit)}
-        >
-          Login
-        </Button>
-      </div>
-    </Form>
   );
 }
